+++ MatrixSDK/MXSession.m
@@ -845,8 +845,20 @@ typedef void (^MXOnResumeDone)(void);
         MXError *mxError = [[MXError alloc] initWithNSError:error];
         if ([mxError.errcode isEqualToString:kMXErrCodeStringUnknownToken])
         {
-            NSLog(@"[MXSession] The access token is no more valid. Go to MXSessionStateUnknownToken state.");
-            [self setState:MXSessionStateUnknownToken];
+            NSLog(@"[MXSession] isUnknownTokenError: The access token is no more valid.");
+
+            if (mxError.httpResponse.statusCode == 401
+                && [mxError.userInfo[kMXErrorSoftLogoutKey] isEqual:@(YES)])
+            {
+                NSLog(@"[MXSession] isUnknownTokenError: Go to MXSessionStateSoftLogout state.");
+                [self setState:MXSessionStateSoftLogout];
+            }
+            else
+            {
+                NSLog(@"[MXSession] isUnknownTokenError: Go to MXSessionStateUnknownToken state.");
+                [self setState:MXSessionStateUnknownToken];
+            }
+
             return YES;
         }
     }